return to average
